Rename parse_gameinfo to parse_game_info
