+++ code/run_tests.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
-python3 -m tests.test_gameinfo
+python3 -m tests.test_game_info
 python3 -m tests.test_team_stats
