+++ code/tests/test_game_info.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import unittest
-from parse_raw_data.parse_gameinfo import convert_time, convert_weather
+from parse_raw_data.parse_game_info import convert_time, convert_weather
 
 
 class TestGameInfo(unittest.TestCase):
@@ -17,19 +17,19 @@ class TestGameInfo(unittest.TestCase):
     def test_convert_weather(self):
         # Successful
         self.assertEqual(
-                convert_weather("13 degrees, relative humidity 62%, wind 8 mph"), 
+                convert_weather("13 degrees, relative humidity 62%, wind 8 mph"),
                 {"temperature": 13, "relative humidity": 0.62, "wind speed": 8, "wind chill": None}
                 )
         self.assertEqual(
-                convert_weather("35 degrees, relative humidity 59%, wind 10 mph, wind chill 27"), 
+                convert_weather("35 degrees, relative humidity 59%, wind 10 mph, wind chill 27"),
                 {"temperature": 35, "relative humidity": 0.59, "wind speed": 10, "wind chill": 27}
                 )
         self.assertEqual(
-                convert_weather("72 degrees, no wind"), 
+                convert_weather("72 degrees, no wind"),
                 {"temperature": 72, "relative humidity": None, "wind speed": 0, "wind chill": None}
                 )
         self.assertEqual(
-                convert_weather("28 degrees, wind 1 mph"), 
+                convert_weather("28 degrees, wind 1 mph"),
                 {"temperature": 28, "relative humidity": None, "wind speed": 1, "wind chill": None}
                 )
         # Failure returns None
